+++ stdlib/build.rs
@@ -1,7 +1,6 @@
 use std::{env, path::Path, sync::Arc};
 
 use assembly::{
-    ast::AstSerdeOptions,
     diagnostics::{IntoDiagnostic, Result},
     library::Library,
     LibraryNamespace, Version,
@@ -36,12 +35,11 @@ fn main() -> Result<()> {
     // write the masl output
     let build_dir = env::var("OUT_DIR").unwrap();
     let build_dir = Path::new(&build_dir);
-    let options = AstSerdeOptions::new(false);
     let output_file = build_dir
         .join(ASL_DIR_PATH)
         .join("std")
         .with_extension(Library::LIBRARY_EXTENSION);
-    stdlib.write_to_file(output_file, options).into_diagnostic()?;
+    stdlib.write_to_file(output_file).into_diagnostic()?;
 
     Ok(())
 }